Consolidate dns implementation.
Similarly to what has happend to the http implentation all the
headers belonging to che dns implementation are grouped in a file.

Same has happend fos the DNS documentation.
