+++ src/dns/resolver.cpp
@@ -2,9 +2,7 @@
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
 
-#include <measurement_kit/dns/defines.hpp>
-#include <measurement_kit/dns/response.hpp>
-#include <measurement_kit/dns/resolver.hpp>
+#include <measurement_kit/dns.hpp>
 
 #include <measurement_kit/common/libs.hpp>
 #include <measurement_kit/common/logger.hpp>
