+++ src/dns/response.cpp
@@ -2,7 +2,7 @@
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
 
-#include <measurement_kit/dns/response.hpp>
+#include <measurement_kit/dns.hpp>
 
 #include "src/common/libs_impl.hpp"
 #include <measurement_kit/common/logger.hpp>
